requirements.txt
dash==2.17.0
dash_bootstrap_components==1.6.0
dash_core_components==2.0.0
dash_html_components==2.0.0
joblib==1.3.2
numpy==1.26.4
opencv_python_headless==4.9.0.80
Pillow==10.3.0
tensorflow==2.16.1
scikit-learn==1.4.1.post1